refactor: Separate UI and domain logic in Race.js
