+++ src/domain/util/random.js
@@ -1,0 +1,6 @@
+export const getRandomNumber = () => {
+  return (
+    Math.floor(Math.random() * (MAX_RANDOM_VALUE - MIN_RANDOM_VALUE + 1)) +
+    MIN_RANDOM_VALUE
+  );
+};
