+++ src/index.js
@@ -1,7 +1,17 @@
-import { startRace } from "./domain/Race.js";
+import { getValidCarNames, getValidAttempt } from "./ui/InputHandler.js";
+import { createCars, moveCars } from "./domain/CarRacing.js";
+import { displayResultTitle, displayRaceResult, displayWinner } from "./ui/OutputHandler.js";
 
 async function run() {
-  await startRace();
+  const carNames = await getValidCarNames();
+  const cars = createCars(carNames);
+  const attempt = await getValidAttempt();
+  displayResultTitle();
+  for (let i = 0; i < attempt; i++) {
+    moveCars(cars);
+    displayRaceResult(cars);
+  }
+  displayWinner(cars);
 }
 
 run();
